fix/connection-manager: always use random port instead of default
If in the quic-p2p config we supply port as unspecified (None), quic-p2p will use
the default port which is set to 443, or 0 if 443 is unavailable. While this behaviour
almost always results in having random ports for clients on *nix operating systems
(since low-number ports are restricted to superusers only), it can result in a situation
when both local vault and a client bind to the same port number (443). We observed this
behaviour on Windows and the solution is to always use a random port for clients.